move data/resources to data/resource
