+++ packages/kuma-gui/src/app/resources/data/index.ts
@@ -1,3 +1,2 @@
 export * from './Resource'
 export * from './ResourceTypeDescriptor'
-export * from './Resources'
